Now more suited for ill conditioned behaviour around the roots. For these cases, even if very close to the actual root, the function has a very high differential and the muller tolerance test using f(z) will fail. Instead the routine tests against the roche starting value, to within a specified value. Also added is a mode to turn all testing off, which is appropriate if some kind of external validation is carried out on the roots. This mode essentially reduces the routine to the 'bugged' behaviour present in the first commit from potapov_interpolation!
Add different modes to allow control of recursion. Will still always recurse when poly order higher than specified max. However can now specify whether to recurse on inaccurate I0 roche or if all internal roots not found. The reason for this is because the routine would often recurse indefinitely and never find the root, which takes much time.

Also, the routine should now be much more robust with regards to returning a warning if all of the know roots have not been found (which is likely if the recursion is turned off and always when root validation is turned off). Three warnings possible: warn_root_check_disabled, warn_inaccurate_roche & warn_could_not_locate_roche_root. The warning is return only if the failure is on the final step, otherwise it's added as a note to the returned status bitfield.

Also, ongoing code rationalisation and improved debug reporting.